LayoutTests/fast/scrolling/mac/stateless-user-scroll-scrollend-expected.txt

PASS windowScrollendEventCount is 1
PASS successfullyParsed is true

TEST COMPLETE
